Refactoring, added java docs
