+++ client/src/main/java/io/pravega/client/stream/notifications/events/ScaleEvent.java
@@ -7,11 +7,15 @@
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.pravega.client.stream.notifications;
+package io.pravega.client.stream.notifications.events;
 
+import io.pravega.client.stream.notifications.Event;
 import lombok.Builder;
 import lombok.Data;
 
+/**
+ * Class to represent a scale event.
+ */
 @Data
 @Builder
 public class ScaleEvent implements Event {
